chore: rename widget components
