+++ src/widgets/LayoutFooter/LayoutFooter.tsx
@@ -1,6 +1,6 @@
-import styles from "./Footer.module.css";
+import styles from "./LayoutFooter.module.css";
 
-export function Footer() {
+export function LayoutFooter() {
   return (
     <footer className={styles.footer}>
       <div className={styles.author}>Created By Elagin Anton</div>
