+++ src/widgets/LayoutHeader/LayoutHeader.tsx
@@ -1,6 +1,6 @@
-import styles from "./Header.module.css";
+import styles from "./LayoutHeader.module.css";
 
-export function Header() {
+export function LayoutHeader() {
   return (
     <header className={styles.header}>
       <h1 className={styles.title}>Post List</h1>
